comments working, click throughs working
